feat(#337): move run_record.py into factory_core package; update test import
