+++ dark-factory/scripts/factory_core/__init__.py
@@ -15,10 +15,16 @@ from .board import (
     STATUS_BACKLOG,
     STATUS_REFINED,
 )
+from .breaker import get_retry_count, increment_retry, reset_retry, trip_to_blocked
+from .deconflict import resolve_merge_conflicts, tier1, tier2, hard_grep_survivors
+from . import run_record
 
 __all__ = [
     "find_board_item", "set_board_status", "post_or_update_comment",
     "OWNER", "REPO", "PROJECT_NUMBER", "PROJECT_ID", "STATUS_FIELD",
     "STATUS_READY", "STATUS_IN_PROGRESS", "STATUS_IN_REVIEW", "STATUS_BLOCKED",
     "STATUS_DONE", "STATUS_BACKLOG", "STATUS_REFINED",
+    "get_retry_count", "increment_retry", "reset_retry", "trip_to_blocked",
+    "resolve_merge_conflicts", "tier1", "tier2", "hard_grep_survivors",
+    "run_record",
 ]
